FIX : layout 폴더구조 변경, 레이아웃 분리작업 필요
