+++ layout/search.style.ts
